basic app structure
This app allows for several different "skins" to be loaded over the
sbemail content. The MainView class allows for switching between the
primary email screen and splash screens such as "flagrant system error"
and "deleted!", as well as the loading of different skins. To accomplish
this, screens can call the `reassert` method when they are ready to
return to the "primary" screen.

sbemail data is stored in a JSON file. On startup, a sbemail will be
started and loaded into the screen. Later down the line, options should
be added to allow for various options in sbemail selection.